requirements.txt
base58==2.1.0
certifi==2020.12.5
chardet==4.0.0
colorzero==2.0
idna==2.10
ipfshttpclient==0.7.0
multiaddr==0.0.9
netaddr==0.8.0
pinatapy-vourhey==0.1.3
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
urllib3==1.26.4
varint==1.0.2
